# openapi.yml
openapi: 3.0.3
info:
  title: Manga Manager API
  description: A self-hosted REST API to manage manga reading progress.
  license:
    name: ''
  version: 1.0.0
paths:
  /manga:
    get:
      tags:
      - handlers::manga
      operationId: list_manga
      parameters:
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int64
          nullable: true
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
          nullable: true
      - name: filter
        in: query
        required: false
        schema:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
          nullable: true
      responses:
        '200':
          description: List manga successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
    post:
      tags:
      - handlers::manga
      operationId: create_manga
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManga'
        required: true
      responses:
        '200':
          description: Manga created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /manga/{id}:
    get:
      tags:
      - handlers::manga
      operationId: get_manga
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Get manga details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Manga not found
      security:
      - bearer_auth: []
    delete:
      tags:
      - handlers::manga
      operationId: delete_manga
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Manga deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Manga not found
      security:
      - bearer_auth: []
    patch:
      tags:
      - handlers::manga
      operationId: update_manga
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManga'
        required: true
      responses:
        '200':
          description: Manga updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Manga not found
      security:
      - bearer_auth: []
  /manga/{id}/history:
    get:
      tags:
      - handlers::manga
      operationId: get_manga_history
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Get manga reading history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /manga/{id}/source:
    get:
      tags:
      - handlers::manga
      operationId: get_manga_sources
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Get manga sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /manga/{id}/source/{domain}:
    delete:
      tags:
      - handlers::manga
      operationId: delete_manga_source
      parameters:
      - name: id
        in: path
        description: Manga ID
        required: true
        schema:
          type: integer
          format: int64
      - name: domain
        in: path
        description: Website domain
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Manga source deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Manga or source not found
      security:
      - bearer_auth: []
  /setting:
    get:
      tags:
      - handlers::setting
      operationId: list_settings
      responses:
        '200':
          description: List all settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /setting/{key}:
    post:
      tags:
      - handlers::setting
      operationId: update_setting
      parameters:
      - name: key
        in: path
        description: Setting key
        required: true
        schema:
          type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Setting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /website:
    get:
      tags:
      - handlers::website
      operationId: list_websites
      responses:
        '200':
          description: List all websites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
  /website/{domain}:
    get:
      tags:
      - handlers::website
      operationId: check_website
      parameters:
      - name: domain
        in: path
        description: Website domain
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Check if website exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
    post:
      tags:
      - handlers::website
      operationId: create_website
      parameters:
      - name: domain
        in: path
        description: Website domain
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Website registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearer_auth: []
components:
  schemas:
    ApiResponse:
      type: object
      required:
      - status
      - message
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/T'
          nullable: true
        message:
          type: string
        status:
          type: string
    Chapter:
      type: object
      required:
      - id
      - manga_id
      - number
      - updated_at
      properties:
        id:
          type: integer
          format: int64
        manga_id:
          type: integer
          format: int64
        number:
          type: string
        updated_at:
          $ref: '#/components/schemas/NaiveDateTime'
    CreateManga:
      type: object
      required:
      - name
      - cover
      - cover_small
      properties:
        cover:
          type: string
        cover_small:
          type: string
        name:
          type: string
        source_path:
          type: string
          nullable: true
        website_domain:
          type: string
          nullable: true
    Existence:
      type: object
      required:
      - existing
      properties:
        existing:
          type: boolean
    HistoryItem:
      type: object
      required:
      - number
      - updated_at
      properties:
        number:
          type: string
        updated_at:
          $ref: '#/components/schemas/chrono.NaiveDateTime'
    Manga:
      type: object
      required:
      - id
      - name
      - cover
      - cover_small
      properties:
        cover:
          type: string
        cover_small:
          type: string
        id:
          type: integer
          format: int64
        name:
          type: string
    MangaDetail:
      type: object
      required:
      - id
      - name
      - cover
      properties:
        cover:
          type: string
        current_chapter:
          type: string
          nullable: true
        id:
          type: integer
          format: int64
        last_read_at:
          allOf:
          - $ref: '#/components/schemas/chrono.NaiveDateTime'
          nullable: true
        name:
          type: string
    MangaListItem:
      type: object
      required:
      - id
      - name
      - cover
      properties:
        cover:
          type: string
        current_chapter:
          type: string
          nullable: true
        id:
          type: integer
          format: int64
        name:
          type: string
    Pagination:
      type: object
      properties:
        filter:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
          nullable: true
        page:
          type: integer
          format: int64
          nullable: true
        size:
          type: integer
          format: int64
          nullable: true
    Setting:
      type: object
      required:
      - key
      - value
      properties:
        key:
          type: string
        value:
          type: string
    Source:
      type: object
      required:
      - id
      - manga_id
      - website_id
      - path
      properties:
        id:
          type: integer
          format: int64
        manga_id:
          type: integer
          format: int64
        path:
          type: string
        website_id:
          type: integer
          format: int64
    UpdateManga:
      type: object
      properties:
        chapter_number:
          type: string
          nullable: true
        cover:
          type: string
          nullable: true
        cover_small:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        source_path:
          type: string
          nullable: true
        website_domain:
          type: string
          nullable: true
    Website:
      type: object
      required:
      - id
      - domain
      properties:
        domain:
          type: string
        id:
          type: integer
          format: int64
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT